move new utils from torch_utils to python_utils
